+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ### 9.27.0
 
 * `[system/sensors]` Added package for collecting sensors data
+* `[strutil]` Added method `Substring` for safe substring extraction
+* `[strutil]` Fixed tests and example for `Substr` method
 * `[strutil]` Code refactoring
 
 ### 9.26.3